dataloom-backend/pyproject.toml
[project]
name = "dataloom-backend"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "sqlmodel>=0.0.22",
    "alembic>=1.14",
    "psycopg2-binary>=2.9",
    "pydantic-settings>=2.0",
    "pandas>=2.1",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.6",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
    "pre-commit>=4.0",
    "httpx>=0.27",
]

[tool.ruff]
line-length = 120
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.ruff.format]
quote-style = "double"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
